backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-chess==1.999
pydantic==2.9.1
pydantic-settings==2.5.2
langchain==0.2.16
langgraph==0.2.14
langchain-groq==0.1.9
pytest==8.3.2
httpx==0.27.2
